Seth's Stuff
I played a bunch. As is, in this project build, the average I was getting was around 15ish depending on my path. So I changed the limit. I added Frank to the prefabs. I added a bit of code that handles objects with the tag "Ceiling" and "Floor" to take up some of the jumps if landed upon.